requirements.txt
playwright==1.50.0
pytest==8.3.4
pytest-playwright==0.6.2
requests==2.32.3
python-dotenv==1.0.1
pytest-html==4.1.1
